update game page
