--- frontend/src/pages/game/GameLobby.module.scss
@@ -1,251 +1,0 @@
-@import "../../styles/config.scss";
-
-.container {
-  min-height: calc(100vh - 10rem);
-  padding: 2rem;
-  border-radius: $rounded-xl;
-  background: $bg-secondary-color;
-
-  .lobbyContainer {
-    max-height: calc(100vh - 10rem);
-    margin: 0;
-    padding: 0;
-    flex-direction: column;
-    display: flex;
-    align-items: left;
-    gap: 2rem;
-    text-align: center;
-    list-style-type: none;
-    overflow-y: auto;
-    scrollbar-width: none; // Firefox
-    -ms-overflow-style: none; // IE (Internet Explorer) and Edge
-
-    &::-webkit-scrollbar {
-      display: none; // WebKit-based browsers like Opera, Chrome, and Safari
-    }
-
-    .createMatch {
-      width: 100%;
-
-      .options {
-        margin-top: 1rem;
-        display: flex;
-        justify-content: space-around;
-        align-items: left;
-
-        form {
-          flex-direction: column;
-          display: flex;
-          align-items: center;
-          gap: 0.25rem;
-
-          label {
-            display: block;
-            font-size: map-get($text-sm, font-size);
-            line-height: map-get($text-sm, line-height);
-            font-weight: $font-medium;
-          }
-
-          input {
-            height: 2rem;
-            padding: 0 1rem 0 1rem;
-            box-sizing: border-box;
-            border: none;
-            border-radius: $rounded;
-            outline: none;
-            background: $extra-tertiary-color;
-            font-size: map-get($text-sm, font-size);
-            line-height: map-get($text-sm, line-height);
-            font-weight: $font-normal;
-            color: $text-primary-color;
-          }
-
-          button {
-            all: unset;
-            margin-top: 0.5rem;
-            padding: 0.5rem 1rem;
-            border-radius: $rounded-lg;
-            transition: background 0.3s ease;
-            font-size: map-get($text-sm, font-size);
-            line-height: map-get($text-sm, line-height);
-            font-weight: $font-medium;
-            color: white;
-            cursor: pointer;
-
-            &.enabled {
-              background: $sucess-primary-color;
-
-              &:hover {
-                background: $sucess-secondary-color;
-              }
-            }
-
-            &.disabled {
-              background: $extra-tertiary-color;
-            }
-          }
-        }
-      }
-    }
-
-    .findMatch {
-      width: 100%;
-
-      .findMatchOptions {
-        margin-top: 1rem;
-        flex-direction: column;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        gap: 1rem;
-
-        .joinQueue {
-          flex-direction: column;
-          display: flex;
-          justify-content: center;
-          align-items: center;
-          gap: 0.25rem;
-
-          label {
-            display: block;
-            font-size: map-get($text-lg, font-size);
-            line-height: map-get($text-lg, line-height);
-            font-weight: $font-medium;
-          }
-
-          button {
-            all: unset;
-            padding: 0.5rem 1rem;
-            border-radius: $rounded-lg;
-            transition: background 0.3s ease;
-            font-size: map-get($text-sm, font-size);
-            line-height: map-get($text-sm, line-height);
-            font-weight: $font-medium;
-            color: white;
-            cursor: pointer;
-
-            &.confirm {
-              background: $sucess-primary-color;
-
-              &:hover {
-                background: $sucess-secondary-color;
-              }
-            }
-
-            &.cancel {
-              background: #762c2d;
-
-              &:hover {
-                background: #da3633;
-              }
-            }
-          }
-        }
-
-        .joinableGames {
-          width: 100%;
-          flex-direction: column;
-          display: flex;
-          justify-content: center;
-          align-items: center;
-
-          label {
-            display: block;
-            font-size: map-get($text-lg, font-size);
-            line-height: map-get($text-lg, line-height);
-            font-weight: $font-medium;
-          }
-
-          p {
-            font-size: map-get($text-sm, font-size);
-            line-height: map-get($text-sm, line-height);
-            font-weight: $font-normal;
-            color: $text-secondary-color;
-          }
-
-          table {
-            width: 100%;
-            text-align: center;
-
-            thead {
-              th {
-                font-size: map-get($text-base, font-size);
-                line-height: map-get($text-base, line-height);
-                font-weight: $font-semibold;
-
-                &:nth-child(1) {
-                  text-align: left;
-                }
-              }
-            }
-
-            tbody {
-              td {
-                padding: 0.5rem 0;
-                font-size: map-get($text-base, font-size);
-                line-height: map-get($text-base, line-height);
-                font-weight: $font-normal;
-                border-bottom: 1px solid rgb(255, 255, 255, 0.05);
-
-                &:nth-child(1) {
-                  max-width: 3rem;
-                  text-align: left;
-                  overflow: hidden;
-                  text-overflow: ellipsis;
-                  white-space: nowrap;
-                }
-
-                button {
-                  all: unset;
-                  padding: 0.5rem 1rem;
-                  border-radius: $rounded-lg;
-                  transition: background 0.3s ease;
-                  font-size: map-get($text-sm, font-size);
-                  line-height: map-get($text-sm, line-height);
-                  font-weight: $font-medium;
-                  color: white;
-                  cursor: pointer;
-
-                  &.enabled {
-                    background: $sucess-primary-color;
-
-                    &:hover {
-                      background: $sucess-secondary-color;
-                    }
-                  }
-
-                  &.disabled {
-                    background: $extra-tertiary-color;
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-}
-
-@media (max-width: 768px) {
-  .container {
-    min-height: calc(100vh - 8rem);
-    padding: 1rem;
-
-    .lobbyContainer {
-      max-height: none;
-      overflow-y: visible;
-      align-items: center;
-      gap: 1rem;
-      text-align: center;
-
-      .createMatch {
-        .options {
-          margin-top: 0rem;
-          flex-direction: column;
-          display: flex;
-        }
-      }
-    }
-  }
-}
