--- frontend/src/pages/game/GamePong.module.scss
@@ -1,44 +1,0 @@
-@import "../../styles/config.scss";
-
-.container {
-  min-height: calc(100vh - 10rem);
-  padding: 2rem;
-  border-radius: $rounded-xl;
-  background: $bg-secondary-color;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  .gameContainer {
-    max-height: calc(100vh - 10rem);
-    margin: 0;
-    padding: 0;
-    list-style-type: none;
-    overflow-y: auto;
-    scrollbar-width: none; // Firefox
-    -ms-overflow-style: none; // IE (Internet Explorer) and Edge
-
-    &::-webkit-scrollbar {
-      display: none; // WebKit-based browsers like Opera, Chrome, and Safari
-    }
-
-    canvas {
-      border-radius: $rounded;
-      border-top: 10px solid #d3d3d3;
-      border-bottom: 10px solid #d3d3d3;
-      outline: none;
-    }
-  }
-}
-
-@media (max-width: 768px) {
-  .container {
-    min-height: calc(100vh - 8rem);
-    padding: 1rem;
-
-    .gameContainer {
-      max-height: none;
-      overflow-y: visible;
-    }
-  }
-}
