--- frontend/src/pages/game/GameWin.module.scss
@@ -1,65 +1,0 @@
-@import "../../styles/config.scss";
-
-.container {
-  min-height: calc(100vh - 10rem);
-  padding: 2rem;
-  border-radius: $rounded-xl;
-  background: $bg-secondary-color;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  .winContainer {
-    max-height: calc(100vh - 10rem);
-    margin: 0;
-    padding: 0;
-    flex-direction: column;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: 1rem;
-    list-style-type: none;
-    overflow-y: auto;
-    scrollbar-width: none; // Firefox
-    -ms-overflow-style: none; // IE (Internet Explorer) and Edge
-
-    &::-webkit-scrollbar {
-      display: none; // WebKit-based browsers like Opera, Chrome, and Safari
-    }
-
-    h1 {
-      font-size: map-get($text-3xl, font-size);
-      line-height: map-get($text-3xl, line-height);
-      font-weight: $font-semibold;
-    }
-
-    button {
-      all: unset;
-      padding: 0.5rem 1rem;
-      border-radius: $rounded-lg;
-      background: $sucess-primary-color;
-      transition: background 0.3s ease;
-      font-size: map-get($text-base, font-size);
-      line-height: map-get($text-base, line-height);
-      font-weight: $font-medium;
-      color: white;
-      cursor: pointer;
-
-      &:hover {
-        background: $sucess-secondary-color;
-      }
-    }
-  }
-}
-
-@media (max-width: 768px) {
-  .container {
-    min-height: calc(100vh - 8rem);
-    padding: 1rem;
-
-    .winContainer {
-      max-height: none;
-      overflow-y: visible;
-    }
-  }
-}
