+++ frontend/src/pages/settings/Settings.tsx
@@ -15,21 +15,19 @@ function Settings() {
   return (
     <>
       {user && (
-        <div className="m-auto max-w-screen-lg">
-          <Tabs defaultValue="profile" className="w-full">
-            <TabsList className="w-full h-full grid grid-cols-1 md:grid-cols-2">
-              <TabsTrigger value="profile">Public profile</TabsTrigger>
-              <TabsTrigger value="auth">Password and authentification</TabsTrigger>
-            </TabsList>
-            <TabsContent value="profile">
-              <ProfileSettings />
-            </TabsContent>
-            <TabsContent value="auth">
-              <AuthSettings />
-              <UnlockUser />
-            </TabsContent>
-          </Tabs>
-        </div>
+        <Tabs defaultValue="profile" className="m-auto max-w-screen-lg">
+          <TabsList className="w-full h-full grid grid-cols-1 md:grid-cols-2">
+            <TabsTrigger value="profile">Public profile</TabsTrigger>
+            <TabsTrigger value="auth">Password and authentification</TabsTrigger>
+          </TabsList>
+          <TabsContent value="profile">
+            <ProfileSettings />
+          </TabsContent>
+          <TabsContent value="auth">
+            <AuthSettings />
+            <UnlockUser />
+          </TabsContent>
+        </Tabs>
       )}
     </>
   );
