+++ frontend/src/styles/globals.scss
@@ -6,24 +6,33 @@
   :root {
     color-scheme: dark;
     --background: #171930;
-    --foreground: 0 0% 98%;
+    --foreground: #e2e8f0;
+
     --card: #25273f;
-    --card-foreground: 0 0% 98%;
+    --card-foreground: #e2e8f0;
+
     --popover: #171930;
-    --popover-foreground: 0 0% 98%;
-    --primary: 0 0% 98%;
+    --popover-foreground: #e2e8f0;
+
+    --primary: #e2e8f0;
     --primary-foreground: 240 5.9% 10%;
+
     --secondary: #374151;
-    --secondary-foreground: 0 0% 98%;
+    --secondary-foreground: #e2e8f0;
+
     --muted: #25273f;
     --muted-foreground: 240 5% 64.9%;
+
     --accent: #25273f;
-    --accent-foreground: 0 0% 98%;
+    --accent-foreground: #e2e8f0;
+
     --destructive: 0 62.8% 30.6%;
-    --destructive-foreground: 0 0% 98%;
+    --destructive-foreground: #e2e8f0;
+
     --border: #25273f;
     --input: #25273f;
     --ring: 240 4.9% 83.9%;
+
     --chart-1: 220 70% 50%;
     --chart-2: 160 60% 45%;
     --chart-3: 30 80% 55%;
